mox-cu/02_web3py_favorites_cu/pyproject.toml
[project]
name = "02-web3py-favorites-cu"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "python-dotenv>=1.0.1",
    "vyper>=0.4.0",
    "web3>=7.6.1",
]
